🚚 [RUMF-1449] move addEventListener functions to a dedicated module
