+++ packages/core/src/domain/report/reportObservable.ts
@@ -1,7 +1,8 @@
 import { toStackTraceString } from '../../tools/error'
 import { monitor } from '../../tools/monitor'
 import { mergeObservables, Observable } from '../../tools/observable'
-import { DOM_EVENT, includes, addEventListener, safeTruncate } from '../../tools/utils'
+import { includes, safeTruncate } from '../../tools/utils'
+import { addEventListener, DOM_EVENT } from '../../browser/addEventListener'
 import type { Report, BrowserWindow, ReportType } from './browser.types'
 
 export const RawReportType = {
