+++ packages/core/src/domain/session/sessionManager.spec.ts
@@ -2,9 +2,10 @@ import type { CookieOptions } from '../../browser/cookie'
 import { COOKIE_ACCESS_DELAY, getCookie, setCookie } from '../../browser/cookie'
 import type { Clock } from '../../../test/specHelper'
 import { mockClock, restorePageVisibility, setPageVisibility, createNewEvent } from '../../../test/specHelper'
-import { ONE_HOUR, DOM_EVENT, ONE_SECOND } from '../../tools/utils'
+import { ONE_HOUR, ONE_SECOND } from '../../tools/utils'
 import type { RelativeTime } from '../../tools/timeUtils'
 import { isIE } from '../../tools/browserDetection'
+import { DOM_EVENT } from '../../browser/addEventListener'
 import type { SessionManager } from './sessionManager'
 import { startSessionManager, stopSessionManager, VISIBILITY_CHECK_DELAY } from './sessionManager'
 import { SESSION_COOKIE_NAME } from './sessionCookieStore'
